feat: finish mobile first 320px
