+++ js/p1.f54cae6aec942be83832.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";var e={225:(e,t,o)=>{o.r(t)},123:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.arrowUp=void 0,t.arrowUp=document.querySelector(".footer_arrow")},56:()=>{const e=function(e){e.style.wordBreak="break-word";const t=e.innerText.split("").map(((e,t)=>{const o=document.createElement("span");return"["!==e&&"]"!==e&&"'"!==e&&"!"!==e||(o.style.color="#E12179"),o.innerHTML=e,o.style.opacity=0,o.style.transitionDelay=.1*t+"s",o}));return e.innerHTML="",e.append(...t),t}(document.querySelector(".header_title_text"));setTimeout((()=>e.forEach((e=>e.style.opacity=1))),0)},395:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0});const o=document.querySelector(".header_menu_btn"),l=document.querySelector(".header_menu"),n=document.querySelector(".header"),r=document.querySelectorAll(".header_menu_line"),i=document.querySelector(".header_info_inner"),s=document.querySelector(".header_video_inner"),c=document.querySelector(".header_social"),u=document.querySelector(".header_menu_inner"),d=()=>{null==l||l.classList.toggle("header_menu_hiden"),null==c||c.classList.toggle("header_menu_hiden"),null==c||c.classList.toggle("header_social_anim"),null==n||n.classList.toggle("header_style"),null==i||i.classList.toggle("header_hide"),null==s||s.classList.toggle("header_hide"),r.forEach(((e,t)=>{e.classList.toggle(`header_menu_line_${t+1}`)}))};null==o||o.addEventListener("click",d),null==u||u.addEventListener("click",(e=>{let t=e.target;const o=document.querySelectorAll(".header_menu_link");for(let e=0;e<o.length;e++)if(t===o[e]){d();break}}))},719:()=>{const e=document.querySelector(".header_video_block"),t=document.querySelector(".video"),o=document.querySelector(".video_main"),l=(document.querySelector(".video_controls"),document.querySelector(".video_btn_close")),n=document.querySelector(".video_control_play"),r=document.querySelector(".video_control_pause"),i=document.querySelector(".video_btn_play"),s=document.querySelector(".video_control_progress"),c=document.querySelector(".video_control_line"),u=document.querySelector(".video_control_marker"),d=document.querySelector(".video_control_sound"),a=document.querySelector(".video_control_volume"),f=document.querySelector(".video_control_full");window.onload=function(){if(!o)throw new Error("Object is possibly 'null'");o.volume=.4};const y=()=>{null==t||t.classList.add("video_hiden"),m(),document.fullscreenElement&&p()},_=()=>{if(null==o||o.play(),!(n&&r&&i))throw new Error("Object is possibly 'null'");n.style.display="none",r.style.display="block",i.style.display="none"},m=()=>{if(null==o||o.pause(),!(r&&n&&i))throw new Error("Object is possibly 'null'");r.style.display="none",n.style.display="block",i.style.display="block"},v=()=>{if(!(c&&o&&s&&u))throw new Error("Object is possibly 'null'");{const e=o.currentTime/o.duration*s.clientWidth;c.style.width=`${e}px`,u.style.transform=`translateX(${e}px)`}(null==o?void 0:o.currentTime)===(null==o?void 0:o.duration)&&m()},p=()=>{document.fullscreenElement?document.exitFullscreen():null==t||t.requestFullscreen()};!function(){if(!o)throw new Error("Object is possibly 'null'");o.ontimeupdate=v}(),null==o||o.addEventListener("click",(()=>{o.paused?_():m()})),null==t||t.addEventListener("click",(e=>{e.target===t&&y()})),null==a||a.addEventListener("input",(function(){const e=this.value;if(o&&a){let t=e/100;t>0?null==d||d.classList.remove("video_control_sound_off"):null==d||d.classList.add("video_control_sound_off"),o.volume=t,a.style.background=`linear-gradient(to right, #ffffff 0%, #ffffff ${100*t}%, rgba(210, 108, 213, 0.5) ${100*t}%, rgba(210, 108, 213, 0.5))`}})),null==d||d.addEventListener("click",(()=>{if(0!==(null==o?void 0:o.volume)&&localStorage.setItem("prevVolume",JSON.stringify(null==o?void 0:o.volume)),o){let e=JSON.parse(localStorage.getItem("prevVolume")||"");0===o.volume?(null==d||d.classList.remove("video_control_sound_off"),a.value=100*e,o.volume=e,a.style.background=`linear-gradient(to right, #ffffff 0%, #ffffff ${a.value}%, rgba(210, 108, 213, 0.5) ${a.value}%, rgba(210, 108, 213, 0.5))`):(null==d||d.classList.add("video_control_sound_off"),o.volume=0,a.value=0,a.style.background="linear-gradient(to right, #ffffff 0%, #ffffff 0%, rgba(210, 108, 213, 0.5) 0%, rgba(210, 108, 213, 0.5))")}})),null==s||s.addEventListener("click",(e=>{let t=null==s?void 0:s.clientWidth,l=e.offsetX;if(e.value=100*l/t,!o||!u)throw new Error("Object is possibly 'null'");o.currentTime=o.duration*(l/t)})),null==i||i.addEventListener("click",_),null==r||r.addEventListener("click",m),null==n||n.addEventListener("click",_),null==l||l.addEventListener("click",y),null==e||e.addEventListener("click",(()=>{null==t||t.classList.remove("video_hiden")})),null==f||f.addEventListener("click",p)},358:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.showAnim=t.startPoints=void 0,t.startPoints=document.querySelectorAll(".anim_0");const o=function(e){let t=[];for(let o=0;o<e.length;o++){e[o].style.wordBreak="break-word";const l=e[o].innerText.split("").map(((e,t)=>{const o=document.createElement("span");return o.innerHTML=e,o.style.opacity="0",o.style.transitionDelay=.3*t+"s",o}));t.push([l]),e[o].innerHTML="",e[o].append(...l)}return t}(document.querySelectorAll(".anim_0"));t.showAnim=function(){o.forEach((e=>{e.forEach((e=>e.forEach((e=>{setTimeout((()=>e.style.opacity="1"))}))))}),0)}},368:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.getOffSet=t.skillsGrafic=void 0,t.skillsGrafic=document.querySelectorAll(".anim"),t.getOffSet=function(e){const t=e.getBoundingClientRect(),o=window.pageXOffset,l=window.pageYOffset;return{top:t.top+l,letf:t.left+o}}},335:(e,t,o)=>{Object.defineProperty(t,"__esModule",{value:!0});const l=o(368),n=o(358),r=o(358),i=o(123);window.addEventListener("scroll",(()=>{n.startPoints.forEach((e=>{const t=e.offsetHeight,o=(0,l.getOffSet)(e).top;let n=window.innerHeight-t;window.pageYOffset>=o-n&&(0,r.showAnim)()})),l.skillsGrafic.forEach((e=>{const t=e.offsetHeight,o=(0,l.getOffSet)(e).top;let n=window.innerHeight-t;window.pageYOffset>=o-n&&e.classList.add("active")})),function(){const e=i.arrowUp.offsetHeight,t=(0,l.getOffSet)(i.arrowUp).top;let o=window.innerHeight-e;window.pageYOffset>t-o&&i.arrowUp.classList.add("show_arrow")}()}))}},t={};function o(l){var n=t[l];if(void 0!==n)return n.exports;var r=t[l]={exports:{}};return e[l](r,r.exports,o),r.exports}o.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o(225),o(395),o(358),o(368),o(719),o(56),o(123),o(335)})();
