+++ src/index.ts
@@ -4,3 +4,5 @@ import "./ts/main/anim_about.ts";
 import "./ts/main/anim_skills";
 import "./ts/header/player";
 import "./ts/header/anim_title";
+import "./ts/footer/arrowUp";
+import "./ts/scroll";
