+++ src/scss/footer.scss
@@ -1,6 +1,7 @@
 .footer {
   padding: 1rem;
   background-color: $color-bg;
+  position: relative;
   &_inner {
     display: flex;
     align-items: center;
@@ -35,5 +36,41 @@
     &:hover {
       opacity: 0.5;
     }
+  }
+  &_arrow {
+    position: absolute;
+    right: 1rem;
+    top: -20rem;
+    background-image: url("../assets/images/arrow-up.png");
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-position: center;
+    width: 4rem;
+    height: 4rem;
+    transition: all 0.5s;
+    animation: arrow 5s linear infinite;
+    text-decoration: none;
+    border: 1px solid $color-main;
+    opacity: 0;
+    transform: translateX(-1000%);
+    overflow: hidden;
+  }
+}
+
+.show_arrow {
+  opacity: 1;
+  transform: translateX(0%);
+  position: absolute;
+}
+
+@keyframes arrow {
+  0% {
+    opacity: 0.5;
+  }
+  50% {
+    opacity: 0.7;
+  }
+  100% {
+    opacity: 1;
   }
 }
