+++ src/scss/header.scss
@@ -132,10 +132,13 @@
       &_item {
         padding-bottom: 1.5rem;
         text-align: center;
+        text-transform: uppercase;
       }
       &_link {
-        font-size: 1.7rem;
+        font-size: 1.4rem;
+        line-height: 1.8rem;
         text-decoration: none;
+        font-weight: 300;
         color: $color-main;
         cursor: pointer;
         transition: all 0.5s;
