+++ src/ts/main/anim_skills.ts
@@ -1,19 +1,4 @@
-const skillsGrafic = document.getElementsByClassName(
-  "anim"
-) as HTMLCollectionOf<HTMLElement>;
-
-function getCor() {
-  for (let i = 0; i < skillsGrafic.length; i++) {
-    let el = skillsGrafic[i];
-    const height = el.offsetHeight;
-    const offSet = getOffSet(el).top;
-    let point = window.innerHeight - height;
-
-    if (window.pageYOffset > offSet - point) {
-      el.classList.add("active");
-    }
-  }
-}
+export const skillsGrafic: any = document.querySelectorAll(".anim");
 
 export function getOffSet(el) {
   const rect = el.getBoundingClientRect();
@@ -25,5 +10,3 @@ export function getOffSet(el) {
     letf: rect.left + scrollLeft,
   };
 }
-
-window.addEventListener("scroll", getCor);
